Remove path clean function
We can assume the search path is input at the last position of
arguments so that if we want to use the clean path function we
should check all arguments and find out the search path then
we clean it, but this is so redundant. Thus, we remove this
function.